Move Thread sources into the clp namespace.
